+++ p2p/tests/connection.rs
@@ -286,7 +286,6 @@ async fn mutual_rust_to_libp2p() -> anyhow::Result<()> {
 
 /// Tests that a Rust node can resolve mutual connection between itself and a libp2p-based node.
 #[tokio::test]
-#[ignore = "TODO: see https://github.com/o1-labs/mina-rust/issues/399"]
 async fn mutual_rust_to_libp2p_port_reuse() -> anyhow::Result<()> {
     let mut cluster = ClusterBuilder::default()
         .ports_with_len(10)
@@ -323,7 +322,7 @@ async fn mutual_rust_to_libp2p_port_reuse() -> anyhow::Result<()> {
     .await?;
     assert!(connected);
 
-    // try_run_cluster(&mut cluster, Duration::from_secs(2)).await?;
+    try_run_cluster(&mut cluster, Duration::from_secs(2)).await?;
 
     assert_peer_is_ready(&cluster, rust_node, peer_id2);