+++ vrf/src/lib.rs
@@ -255,7 +255,6 @@ mod test {
     }
 
     #[test]
-    #[ignore]
     fn test_slot_calculation_time_big_producer() {
         let start = redux::Instant::now();
         for i in 1..14403 {
@@ -285,7 +284,6 @@ mod test {
     }
 
     #[test]
-    #[ignore]
     fn test_first_winning_slot() {
         for i in 0..7000 {
             let vrf_input = VrfEvaluationInput {